env: build separated bundles for ie11
