+++ apps/calendar/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@toast-ui/calendar",
   "author": "NHN FE Development Lab <dl_javascript@nhn.com>",
-  "version": "2.0.0-dev",
+  "version": "2.0.0-alpha",
   "main": "dist/toastui-calendar.js",
   "types": "types/index.d.ts",
   "sideEffects": [
@@ -48,14 +48,13 @@
     "check-types": "tsc --project ./tsconfig.json --noEmit",
     "validate": "npm run check-types && npm run lint",
     "lint": "eslint .",
-    "serve": "webpack serve --config webpack.dev.config.js",
     "cpy-dist2doc": "mkdir -p doc/dist && cp -f -r dist doc",
     "release-note": "tuie",
-    "build": "rm -rf dist && npm run build:es5 && npm run build:es6 && npm run build:esm",
-    "build:es5": "webpack && webpack --env minify",
-    "build:es6": "webpack --config webpack.es6.config.js && webpack --config webpack.es6.config.js --env minify",
+    "build:all": "rm -rf dist && npm run build && npm run build:ie11",
+    "build": "webpack --config webpack.config.js && webpack --config webpack.config.js --env minify",
+    "build:ie11": "webpack --config webpack.config.js --env ie11 && webpack --config webpack.config.js --env minify ie11",
     "build:esm": "tsc -p tsconfig.esm.json && prettier dist/esm/**/*.js --write",
-    "analyze": "webpack --minify --profile --json > stats.json && webpack-bundle-analyzer stats.json ./dist",
+    "analyze": "webpack --config webpack.config.js --env --profile --json > stats.json && webpack-bundle-analyzer stats.json ./dist",
     "storybook": "start-storybook -p 6006",
     "storybook:build": "build-storybook",
     "storybook:build:docs": "STORYBOOK_ENV=docs build-storybook",
