+++ babel.config.json
@@ -1,14 +1,6 @@
 {
   "presets": [
-    [
-      "@babel/preset-env",
-      {
-        "useBuiltIns": "usage",
-        "corejs": 3,
-        "shippedProposals": true,
-        "targets": "defaults"
-      }
-    ],
+    ["@babel/preset-env"],
     ["@babel/preset-typescript", { "jsxPragma": "h" }]
   ],
   "plugins": [["@babel/plugin-transform-react-jsx", { "pragma": "h", "pragmaFrag": "Fragment" }]]
